requirements.txt
# Air NZ AI Governance Platform - Dependencies

# Core dependencies
python-dateutil>=2.8.2
typing-extensions>=4.5.0

# OpenAI API (LLM service)
openai>=1.3.0

# HTTP requests (for Flight API)
requests>=2.31.0

# Python environment variables
python-dotenv>=1.0.0

# Optional production dependencies
# anthropic>=0.18.0         # Claude API alternative
# langchain>=0.1.0          # LangChain orchestration
# chromadb>=0.4.0           # Vector store for RAG
# redis>=5.0.0              # Caching and session management
# sqlalchemy>=2.0.0         # Database ORM (for PostgreSQL migration)
# pydantic>=2.0.0           # Data validation
# fastapi>=0.109.0          # API framework
# uvicorn>=0.27.0           # ASGI server
# psycopg2-binary>=2.9.9    # PostgreSQL adapter

# Development dependencies
pytest>=7.4.0
pytest-cov>=4.1.0
black>=23.12.0
flake8>=7.0.0
mypy>=1.8.0

# Documentation
mkdocs>=1.5.3
mkdocs-material>=9.5.3
